+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-## Unreleased
+## 1.0.0
 
 - New config: `depend` via [`eslint-plugin-depend`](https://www.npmjs.com/package/eslint-plugin-depend), **<u>disabled</u>** by default.
 - New config: `math` via [`eslint-plugin-math`](https://www.npmjs.com/package/eslint-plugin-math), enabled by default.
+- [**BREAKING**] You can now specify the method of disabling autofix: `rules-copy` would create a plugin named `disable-autofix` and copy the rules for which the autofix is disabled. `plugin-copy` would copy an entire plugin and replace the origin plugin with the copy. The default method is now `plugin-copy`, which leaves full rule names unchanged. You can control the overall or per-plugin autofix disabling method via the `disableAutofixMethod` option.
+- [**BREAKING**] Removed the ability to disable rule autofix by enabling `disable-autofix/*` rules. Instead, when configuring a rule, you can now return an object and set `disableAutofix: true` in it.
 
 ## 0.10.0